observe zone to maintain in play card state
This could probably be made more generic instead of
being specific to the “in-play” zone.

For example, could have each zone have a zone state
structure of cards that maintain state relevant to
the zone.

Expose methods like:
* new_card?
* reset!
* stuff that delegates to card

Going to stew on it for a bit since I don’t need it
right now.