Push the perfoming of card abilities into the zone
This shifts responsibility from the action to the zone. Now the action
is simply responsible for performing the zone transfer into the "in
play" zone. This is convenient because the in play zone maintains a
shadow collection of "in play state" that it can use to determine what
abilities are performed automatically as a result of the player telling
the game that they would like to play a card.

There was a little bit of weirdness for cards that become other cards,
such as stealth needle. I decided to handle this by adding a small
amount of interface to zone. The method Zone#reset!(card) is a hook that
instructs the zone to re-evaluate the in play state after an ability
like copy ship that effectively replays the card.