move state machine responsibility out of zone
After talking through this, it seems like a breach of responsibility.

The zone’s execute method never paused for player interaction, so it seems its kind of abuse of the state machine.

However, I can see in a more complicated game a scenario where there is player interaction that needs to be resolved before a card finished being played. In that case I think there’s actually a transient zone in between playing the card and the card resolving.